requirements.txt
# These are the libraries that were used to complete the project

aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
annotated-types==0.7.0
ansicolors==1.1.8
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
deep-sort-realtime==1.3.2
defusedxml==0.7.1
entrypoints==0.4
filelock==3.19.1
filetype==1.2.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.9.0
graphviz==0.21
idna==3.7
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lap==0.5.12
lightgbm==4.6.0
llvmlite==0.46.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1761214490209/work
mpmath==1.3.0
multidict==6.7.0
narwhals==2.13.0
nbclient==0.10.2
nbformat==5.10.4
networkx==3.5
numba==0.63.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
pandas==2.3.3
papermill==2.6.0
pi_heif==1.1.1
pillow==11.3.0
pillow-avif-plugin==1.5.2
plotly==6.5.0
propcache==0.4.1
protobuf==6.33.2
pydantic==2.12.5
pydantic_core==2.41.5
pynndescent==0.5.13
pyparsing==3.2.5
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
roboflow==1.2.11
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
setuptools==80.9.0
simsimd==6.5.3
stringzilla==4.4.0
supervision==0.27.0
sympy==1.13.1
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.1
transformers==4.57.3
triton==3.2.0
tzdata==2025.2
ultralytics==8.3.234
ultralytics-thop==2.0.18
umap-learn==0.5.9.post2
urllib3==2.5.0
wheel==0.45.1
yarl==1.22.0

